+++ test/cmdlineTests/message_format_utf8/input.sol
@@ -1,11 +1,11 @@
 contract Foo {
-/* ©©©©ﾩ©©©©© 2017 */ constructor () { "©©©©ﾩ©©©©©" ; }
+/* ©©©©ﾩ©©©©© 2017 */ constructor () { unicode"©©©©ﾩ©©©©©" ; }
 
     function f() public pure {
 
-                        "S = π × r²";
-/* ₀₁₂₃₄⁵⁶⁷⁸⁹ */		"∑ 1/n! ≈ 2.7";                 // tabs in-between
-/* Ŀŏŗėɯ ïƥŝʉɱ */		"μὴ		χεῖρον		βέλτιστον"; // tabs in-between and inside
+                        unicode"S = π × r²";
+/* ₀₁₂₃₄⁵⁶⁷⁸⁹ */		unicode"∑ 1/n! ≈ 2.7";                 // tabs in-between
+/* Ŀŏŗėɯ ïƥŝʉɱ */		unicode"μὴ		χεῖρον		βέλτιστον"; // tabs in-between and inside
 
     }