+++ test/libsolidity/SolidityScanner.cpp
@@ -81,6 +81,45 @@ BOOST_AUTO_TEST_CASE(assembly_multiple_assign)
 	BOOST_CHECK_EQUAL(scanner.next(), Token::EOS);
 }
 
+BOOST_AUTO_TEST_CASE(string_printable)
+{
+	for (unsigned v = 0x20; v < 0x7e; v++) {
+		string lit{static_cast<char>(v)};
+		// Escape \ and " (since we are quoting with ")
+		if (v == '\\' || v == '"')
+			lit = string{'\\'} + lit;
+		Scanner scanner(CharStream("  { \"" + lit + "\"", ""));
+		BOOST_CHECK_EQUAL(scanner.currentToken(), Token::LBrace);
+		BOOST_CHECK_EQUAL(scanner.next(), Token::StringLiteral);
+		BOOST_CHECK_EQUAL(scanner.currentLiteral(), string{static_cast<char>(v)});
+		BOOST_CHECK_EQUAL(scanner.next(), Token::EOS);
+	}
+	// Special case of unescaped " for strings quoted with '
+	Scanner scanner(CharStream("  { '\"'", ""));
+	BOOST_CHECK_EQUAL(scanner.currentToken(), Token::LBrace);
+	BOOST_CHECK_EQUAL(scanner.next(), Token::StringLiteral);
+	BOOST_CHECK_EQUAL(scanner.currentLiteral(), "\"");
+	BOOST_CHECK_EQUAL(scanner.next(), Token::EOS);
+}
+
+BOOST_AUTO_TEST_CASE(string_nonprintable)
+{
+	for (unsigned v = 0; v < 0xff; v++) {
+		// Skip the valid ones
+		if (v >= 0x20 && v <= 0x7e)
+			continue;
+		string lit{static_cast<char>(v)};
+		Scanner scanner(CharStream("  { \"" + lit + "\"", ""));
+		BOOST_CHECK_EQUAL(scanner.currentToken(), Token::LBrace);
+		BOOST_CHECK_EQUAL(scanner.next(), Token::Illegal);
+		if (v == '\n' || v == '\v' || v == '\f' || v == '\r')
+			BOOST_CHECK_EQUAL(scanner.currentError(), ScannerError::IllegalStringEndQuote);
+		else
+			BOOST_CHECK_EQUAL(scanner.currentError(), ScannerError::IllegalCharacterInString);
+		BOOST_CHECK_EQUAL(scanner.currentLiteral(), "");
+	}
+}
+
 BOOST_AUTO_TEST_CASE(string_escapes)
 {
 	Scanner scanner(CharStream("  { \"a\\x61\"", ""));
@@ -506,6 +545,8 @@ BOOST_AUTO_TEST_CASE(empty_comment)
 
 }
 
+// Unicode string escapes
+
 BOOST_AUTO_TEST_CASE(valid_unicode_string_escape)
 {
 	Scanner scanner(CharStream("{ \"\\u00DAnicode\"", ""));
@@ -543,6 +584,25 @@ BOOST_AUTO_TEST_CASE(invalid_short_unicode_string_escape)
 	Scanner scanner(CharStream("{ \"\\uFFnicode\"", ""));
 	BOOST_CHECK_EQUAL(scanner.currentToken(), Token::LBrace);
 	BOOST_CHECK_EQUAL(scanner.next(), Token::Illegal);
+}
+
+// Unicode string literal
+
+BOOST_AUTO_TEST_CASE(valid_unicode_literal)
+{
+	Scanner scanner(CharStream("{ unicode\"Hello 😃\"", ""));
+	BOOST_CHECK_EQUAL(scanner.currentToken(), Token::LBrace);
+	BOOST_CHECK_EQUAL(scanner.next(), Token::UnicodeStringLiteral);
+	BOOST_CHECK_EQUAL(scanner.currentLiteral(), std::string("Hello \xf0\x9f\x98\x83", 10));
+}
+
+BOOST_AUTO_TEST_CASE(valid_nonprintable_in_unicode_literal)
+{
+	// Non-printable characters are allowed in unicode strings...
+	Scanner scanner(CharStream("{ unicode\"Hello \007😃\"", ""));
+	BOOST_CHECK_EQUAL(scanner.currentToken(), Token::LBrace);
+	BOOST_CHECK_EQUAL(scanner.next(), Token::UnicodeStringLiteral);
+	BOOST_CHECK_EQUAL(scanner.currentLiteral(), std::string("Hello \x07\xf0\x9f\x98\x83", 11));
 }
 
 //  HEX STRING LITERAL