+++ test/libsolidity/semanticTests/types/strings.sol
@@ -1,4 +1,7 @@
 contract test {
+    function fixedBytesHex() public returns(bytes32 ret) {
+        return hex"aabb00ff";
+    }
     function fixedBytes() public returns(bytes32 ret) {
         return "abc\x00\xff__";
     }
@@ -11,5 +14,6 @@ contract test {
 // ====
 // compileViaYul: also
 // ----
+// fixedBytesHex() -> "\xaa\xbb\0\xff"
 // fixedBytes() -> "abc\0\xff__"
 // pipeThrough(bytes2, bool): "\0\x02", true -> "\0\x2", true